+++ client/src/pages/ResumeCheck.js
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useState, useEffect } from "react";
 import Container from 'react-bootstrap/Container';
 import ProductList from '../components/checkoutProducts'
 import ProductResume from '../components/checkoutProductsResume'
@@ -7,13 +7,33 @@ import Footer from '../components/footercustomercheckout'
 
 
 function PDP() {
+
+  const [localStorageProducts, setlocalStorageProducts] = useState([]);
+  let localProducts = [];
+
+  useEffect(() => {
+    getLocalStoragePdts();
+  }, []);
+
+  const getLocalStoragePdts = () => {
+    // console.log('localStorage.length: ', localStorage.length);
+    for (let i = 0; i < localStorage.length; i++) {
+      let id = localStorage.key(i);
+      let product = JSON.parse(localStorage.getItem(id));
+      localProducts.push(product);
+      // console.log('productName: ', product.productName);
+    }
+    // console.log('localProducts: ', localProducts);
+    setlocalStorageProducts(localProducts);
+  }
+
   return (
     <>
       <NavCustomer />
       <Container fluid>
         <h2>Review your shopping cart: </h2>
-        <ProductList />
-        <ProductResume />
+        <ProductList localStorageProducts={localStorageProducts} />
+        <ProductResume localStorageProducts={localStorageProducts} />
       </Container>
       <Footer />
     </>
